+++ src/main/scala/Characters/Paladin.scala
@@ -13,7 +13,7 @@ import Armas.Weapon
  * @constructor Creates a new Paladin with the specified attributes.
  *
  */
-class Paladin(name: String, life: Int, defense: Int, weight: Int, weapon: Option[Weapon])
+class Paladin(name: String, life: Int, defense: Int, weight: Double, weapon: Option[Weapon])
   extends AbstractCharacter(name, life, defense, weight, weapon) {
   /** Indicates whether it's the paladin's turn in combat. */
   var isMyTurn: Boolean = false
