+++ src/main/scala/Characters/Warrior.scala
@@ -13,7 +13,7 @@ import Armas.Weapon
  * @constructor Creates a new Geurrero with the specified attributes.
  *
  */
-class Warrior(name: String, life: Int, defense: Int, weight: Int, weapon: Option[Weapon])
+class Warrior(name: String, life: Int, defense: Int, weight: Double, weapon: Option[Weapon])
   extends AbstractCharacter(name, life, defense, weight, weapon) {
 
   /** Indicates whether it's the Warrior's turn in combat. */
