+++ src/main/scala/Characters/WhiteMage.scala
@@ -13,7 +13,7 @@ import Armas.{MagicWeapon, Weapon}
  * @constructor Creates a new White Mage with the specified attributes.
  *
  */
-class WhiteMage(name: String, life: Int, defense: Int, weight: Int, mana: Int, weapon: Option[MagicWeapon])
+class WhiteMage(name: String, life: Int, defense: Int, weight: Double, mana: Int, weapon: Option[MagicWeapon])
   extends AbstractMagicCharacter(name, life, defense, weight, mana, weapon){
 
   /** Indicates whether it's the White Mage's turn in combat. */
